Patched up Swofty's patch up of my implementation of Items in the shop
